It's time to visit the RBAP history book!
I have paid a visit to the RBAP history book and found some great treasures: Many old update logs have been recovered (from a system that exists in older versions of RBAP)! For more info read the added notes to the list of notes about the update log system.

I did this all manually over the span of a few hours so I hope it intrigues someone reading it.